requirements.txt
# Core dependencies
python>=3.8,<3.12
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# LLM/API
openai>=1.0.0
google-generativeai>=0.3.0

# Reinforcement Learning
stable-baselines3>=2.0.0
gymnasium>=0.28.0

# Email parsing
email-validator>=2.0.0
dnspython>=2.4.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter (for analysis)
jupyter>=1.0.0
ipykernel>=6.25.0


